refactor(mesh/media): remove typo and debug code
