+++ packages/mesh/app/actions/get-most-picked-stories-in-category.ts
@@ -4,7 +4,7 @@ import { STATIC_FILE_ENDPOINTS } from '@/constants/config'
 import fetchStatic from '@/utils/fetch-static'
 import { getLogTraceObjectFromHeaders } from '@/utils/log'
 
-import { type Story } from './get-latest-stories-in-categroy'
+import { type Story } from './get-latest-stories-in-category'
 
 export default async function getMostPickedStoriesInCategory(
   categorySlug: string
